refactor: rename the file format to snack_case
