--- src/modules/auth/ForgotPassword/constants/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./ForgotPassword.constants";
-export * from "./ForgotPassword.constants";
