+++ src/modules/auth/ForgotPassword/forgot-password.style.tsx
@@ -1,7 +1,10 @@
 import { styled } from "@mui/material";
 import { Link as LinkRouterDom } from "react-router-dom";
-import { ILinkStyle } from "./types";
-    
+
+export interface ILinkStyle {
+	disabled: boolean;
+}
+
 export const Title = styled("div")(
 	({ theme }) => `
         margin: ${theme.spacing(8, 0, 5)};
@@ -9,7 +12,7 @@ export const Title = styled("div")(
             margin: 0;
             color: ${theme.palette.grey[700]}
         }
-    `   
+    `
 );
 
 export const Footer = styled("footer")(
@@ -22,7 +25,7 @@ export const Footer = styled("footer")(
         hr {
             width: 50%
         }
-    `,
+    `
 );
 
 export const Link = styled(LinkRouterDom)<ILinkStyle>(
