--- src/modules/auth/ForgotPassword/schema/ForgotPassword.schema.ts
@@ -1,5 +1,0 @@
-import * as Yup from "yup";
-
-export const ForgotPasswordSchema = Yup.object({
-	email: Yup.string().email().required(),
-});
