--- src/modules/auth/ForgotPassword/schema/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./ForgotPassword.schema";
-export * from "./ForgotPassword.schema";
