--- src/modules/auth/ForgotPassword/services/types/fetchPostForgotPasswordService.types.ts
@@ -1,5 +1,0 @@
-import { IForgotPasswordForm } from "modules/auth/ForgotPassword/types";
-
-export interface IFetchPostForgotPasswordService {
-	body: IForgotPasswordForm;
-}
