+++ src/modules/auth/ForgotPassword/services/types/post-forgot-password.types.ts
@@ -1,0 +1,5 @@
+import type { IForgotPasswordForm } from "modules/auth/ForgotPassword/types";
+
+export interface IPostForgotPasswordService {
+	body: IForgotPasswordForm;
+}
