+++ src/modules/auth/ForgotPassword/types/index.ts
@@ -1,2 +1,1 @@
-export * as default from "./ForgotPassword.types";
-export * from "./ForgotPassword.types";
+export * from "./forgot-password.types";
